requirements.txt
django==2.2.15
django-nested-admin
markdown2>=2.3.8
MarkupSafe>=1.1.1
mod_wsgi>=4.6.7
mysql-connector>=2.2.9
mysqlclient>=1.4.2.post1
Pillow>=6.1.0
requests>=2.22.0
